+++ src/sll/api/file_api.c
@@ -8,6 +8,7 @@
 #include <sll/handle.h>
 #include <sll/static_object.h>
 #include <sll/stream.h>
+#include <sll/string.h>
 #include <sll/types.h>
 #include <stdint.h>
 #include <stdio.h>
@@ -17,7 +18,7 @@
 
 
 typedef struct __FILE{
-	const sll_char_t nm[SLL_API_MAX_FILE_PATH_LENGTH+1];
+	const sll_char_t nm[SLL_STRING_ALIGN_LENGTH(SLL_API_MAX_FILE_PATH_LENGTH)];
 	sll_string_length_t nml;
 	sll_string_checksum_t nmc;
 	FILE* h;
@@ -121,8 +122,8 @@ __API_FUNC(file_open){
 	_file_fl=tmp;
 _found_index:;
 	file_t* n=malloc(sizeof(file_t));
-	memcpy((sll_char_t*)(n->nm),a->v,a->l*sizeof(sll_char_t));
-	((sll_char_t*)(n->nm))[a->l]=0;
+	memcpy((void*)(n->nm),a->v,a->l*sizeof(sll_char_t));
+	SLL_CHAR(n->nm)[a->l]=0;
 	n->nml=a->l;
 	n->nmc=a->c;
 	n->h=h;