+++ src/sll/string.c
@@ -12,7 +12,7 @@ __SLL_FUNC void sll_string_create(sll_string_length_t l,sll_string_t* o){
 	o->l=l;
 	o->c=0;
 	o->v=malloc(SLL_STRING_ALIGN_LENGTH(l)*sizeof(sll_char_t));
-	o->v[l]=0;
+	SLL_STRING_FORMAT_PADDING(o->v,l);
 }
 
 
@@ -22,7 +22,7 @@ __SLL_FUNC void sll_string_create_from_pointer(const sll_char_t* s,sll_string_t*
 	o->l=l;
 	o->v=malloc(SLL_STRING_ALIGN_LENGTH(l)*sizeof(sll_char_t));
 	memcpy(o->v,s,l);
-	o->v[l]=0;
+	SLL_STRING_FORMAT_PADDING(o->v,l);
 	sll_string_hash(o);
 }
 
@@ -33,16 +33,29 @@ __SLL_FUNC void sll_string_clone(const sll_string_t* s,sll_string_t* d){
 	d->c=s->c;
 	d->v=malloc(SLL_STRING_ALIGN_LENGTH(s->l)*sizeof(sll_char_t));
 	memcpy(d->v,s->v,s->l+1);
+	SLL_STRING_FORMAT_PADDING(d->v,d->l);
 }
 
 
 
 __SLL_FUNC void sll_string_hash(sll_string_t* s){
-	sll_string_checksum_t c=0;
-	for (sll_string_length_t i=0;i<s->l;i++){
-		c^=s->v[i];
+	uint64_t c=0;
+	const uint64_t* p=(const uint64_t*)(s->v);
+	for (sll_string_length_t i=0;i<((s->l+7)>>3);i++){
+		c^=*(p+i);
 	}
-	s->c=c;
+	s->c=((sll_string_checksum_t)c)^((sll_string_checksum_t)(c>>32));
+}
+
+
+
+__SLL_FUNC void sll_string_join(const sll_string_t* a,const sll_string_t* b,sll_string_t* o){
+	o->l=a->l+b->l;
+	o->v=malloc(SLL_STRING_ALIGN_LENGTH(o->l)*sizeof(sll_char_t));
+	SLL_STRING_FORMAT_PADDING(o->v,o->l);
+	memcpy(o->v,a->v,a->l);
+	memcpy(o->v+a->l,b->v,b->l);
+	o->c=SLL_STRING_COMBINE_CHECKSUMS_FAST(a->c,a->l,b->c);
 }