+++ src/sll/url.c
@@ -125,17 +125,15 @@ __SLL_FUNC __SLL_CHECK_OUTPUT sll_return_code_t sll_url_http_request(const sll_s
 	i++;
 	if (o->rc){
 		sll_string_length_t j=i;
-		sll_string_checksum_t c=0;
 		while (r.v[i]!='\r'||r.v[i+1]!='\n'){
 			if (!r.v[i]){
 				goto _error;
 			}
-			c^=r.v[i];
 			i++;
 		}
 		sll_string_create(i-j,o->rc);
-		o->rc->c=c;
 		memcpy(o->rc->v,r.v+j,i-j);
+		sll_string_hash(o->rc);
 	}
 	else{
 		while (r.v[i]!='\r'||r.v[i+1]!='\n'){